+++ forui/test/src/widgets/header/root_header_golden_test.dart
@@ -1,7 +1,7 @@
 @Tags(['golden'])
 library;
 
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
@@ -12,16 +12,16 @@ const title =
     'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.';
 
 void main() {
-  group('FHeader', () {
+  group('FRootHeader', () {
     for (final (name, theme, _) in TestScaffold.themes) {
-      testWidgets('$name with FHeader actions', (tester) async {
+      testWidgets('$name with FRootHeader actions', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
             data: theme,
             child: Padding(
               padding: const EdgeInsets.symmetric(horizontal: 20),
               child: FHeader(
-                title: title,
+                title: const Text(title),
                 actions: [
                   FHeaderAction(
                     icon: FAssets.icons.alarmClock,
@@ -39,36 +39,7 @@ void main() {
 
         await expectLater(
           find.byType(TestScaffold),
-          matchesGoldenFile('header/$name-header.png'),
-        );
-      });
-
-      testWidgets('$name with raw title', (tester) async {
-        await tester.pumpWidget(
-          TestScaffold(
-            data: theme,
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: FHeader(
-                rawTitle: const Text('Title'),
-                actions: [
-                  FHeaderAction(
-                    icon: FAssets.icons.alarmClock,
-                    onPress: null,
-                  ),
-                  FHeaderAction(
-                    icon: FAssets.icons.plus,
-                    onPress: () {},
-                  ),
-                ],
-              ),
-            ),
-          ),
-        );
-
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('header/$name-raw-title.png'),
+          matchesGoldenFile('header/root/$name-header.png'),
         );
       });
     }