+++ samples/lib/main.dart
@@ -60,7 +60,15 @@ class _AppRouter extends $_AppRouter {
         ),
         AutoRoute(
           path: '/header/default',
-          page: HeaderRoute.page,
+          page: RootHeaderRoute.page,
+        ),
+        AutoRoute(
+          path: '/header/nested',
+          page: NestedHeaderRoute.page,
+        ),
+        AutoRoute(
+          path: '/header/nested-x',
+          page: XNestedHeaderRoute.page,
         ),
         AutoRoute(
           path: '/tabs/default',