+++ forui/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0
+
+* Add `Header.nested` widget.
+
+### Breaking changes
+
+* `FHeaderStyle` have been nested in `FHeaderStyles.rootStyle`.
+* `FHeaderActionStyle.action` parameter has been renamed to `FRootHeaderStyle.actionStyle`.
+* `FHeaderActionStyle.padding` parameter has been moved to `FRootHeaderStyle.actionSpacing`.
+
 ## 0.1.0
 
 * Initial release! 🚀